Introduce declaration signature and declaration ID uniqueness.
The declaration engine now tries to keep sure we do not create new
declaration ids for previously added types.

Right now this is only implemented for enums, but should be extended in
the future for other declaration types. This is pretty much the same
idea as the monormophization cache we previously discussed, but works
for non-generic types as well.